Set span attributes from request
Creation of a custom context can be side effecting. The introduction
of the resource Kleisli and subsequent changes meant that if the
creation of a custom context failed, http4s request headers would
not be set. This ensures that the default Kleisli implementation
of `Request => Span` sets span attributes before creating the
custom context.

The http4s tracer still sets the request headers as attributes, in
case someone wants to provide their own `ResourceKleisli` without
setting the request header attributes.