Fix shutdown of QueuedSpanCompleter
By having an uncancelable fiber via background and discarding the reference to it, the shut down wasn't awaiting the sending process to finish, causing spans to be lost. This bug probably got amplified in newer CE versions with improved performance, it manifested regularly in the Jaeger-based tests, see https://github.com/trace4cats/trace4cats-jaeger/pull/175.

I tried adding a test case that would reproduce the issue, but failed as it seems to require real concurrency with thread switches that are hard to set up in a simple test case where as they come naturally from network I/O in integration tests.